requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-core
cookiecutter==1.7.0
azure-core
azure-ai-textanalytics==5.3.0b1
